Remove `Base.lib{blas,lapack}_name`
These have been deprecated for a long time, users should instead use
`LinearAlgebra.BLAS.libblas` or `LinearAlgebra.BLAS.liblapack`.

This will be necessary once we have lazy JLLs, as it's not desirable to
recreate the dependency graph in `Base`, we already have it in standard
libraries, and only load libblastrampoline when `LinearAlgebra` starts
doing ccalls.